lrdb_back/requirements.txt
django == 3.0.5
djangorestframework == 3.11.0
django-cors-headers == 3.3.0
drf-oidc-auth == 0.10.0